+++ plugins/backend/catalog-backend-module-aws/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage/plugin-catalog-backend-module-aws
 
+## 5.5.1
+
+### Patch Changes
+
+- a83e7e9: Tidy unused plugin depdendencies.
+
 ## 5.5.0
 
 ### Minor Changes