+++ plugins/backend/rag-ai-backend/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @roadiehq/rag-ai-backend
 
+## 1.3.3
+
+### Patch Changes
+
+- a83e7e9: Tidy unused plugin depdendencies.
+- Updated dependencies [a83e7e9]
+  - @roadiehq/rag-ai-node@0.1.10
+
+## 1.3.2
+
+### Patch Changes
+
+- dffd381: Add error logging to query endpoint
+
 ## 1.3.1
 
 ### Patch Changes