+++ plugins/frontend/backstage-plugin-launchdarkly/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @roadiehq/backstage-plugin-launchdarkly
 
+## 0.0.11
+
+### Patch Changes
+
+- a83e7e9: Tidy unused plugin depdendencies.
+
 ## 0.0.10
 
 ### Patch Changes