+++ utils/roadie-backstage-entity-validator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @roadiehq/roadie-backstage-entity-validator
 
+## 2.4.14
+
+### Patch Changes
+
+- a83e7e9: Tidy unused plugin depdendencies.
+
 ## 2.4.13
 
 ### Patch Changes